refactor(oci): move digest verification to mvm-contract (301 B1 tier 2)
`verify_sha256_digest` is the "verify" in inspect-and-verify: given bytes
already in hand and the digest they were pinned under, prove they still hash
to it. No registry, no filesystem, no async — so it belongs on the browser
side of the split, and a browser can now check an image it obtained by any
means.

Its whole test module moves with it, which is the point: the seven rejection
cases (tampered content, wrong algorithm, missing prefix, wrong hex length,
uppercase hex, plus the self-consistency guard on the fixture digest) now run
under wasm rather than only on the host. wasm lib tests 554 → 561.

Reuses `verify::encode_hex32`, this crate's existing lowercase-hex encoder,
rather than pulling `hex` in — a browser bundle should not carry a hex crate
for one call. That is why the module is gated on the `protocol` feature,
which is where that helper lives; `protocol` is a default feature, so the
wasm build gets digest verification regardless.

`mvm-fs` re-exports it from the same path as before, so no call site moves.

Verified: wasm32-unknown-unknown builds, 561 wasm lib tests pass, 871 mvm-fs
+ mvm-contract tests pass, `cargo check -p mvm-fs` clean.

Deferred to a later tier: `LinuxPlatform` and `matches_linux_platform` are
pure and belong down here too, but `LinuxPlatform::for_current_arch` reads
`std::env::consts::ARCH`, and moving the type without it means rewriting an
inherent method into a free function across 14 call sites in mvm-cli and
mvm-client. That churn wants its own commit, not this one.
