+++ crates/mvm-fs/src/oci/mod.rs
@@ -65,9 +65,8 @@ pub use archive::{
 };
 pub use error::OciError;
 pub use layer::{LayerDescriptor, LayerFetchOptions, OciLayerFetcher};
-pub use manifest::{
-    FetchedManifest, LinuxPlatform, ManifestFetcher, OciManifestFetcher, verify_sha256_digest,
-};
+pub use manifest::{FetchedManifest, LinuxPlatform, ManifestFetcher, OciManifestFetcher};
+pub use mvm_contract::oci::verify_sha256_digest;
 pub use reference::ImageReference;
 pub use registry::{ClientConfig, ClientProtocol, RegistryAuthConfig};
 pub use unpack::{
